Updated the next value code to add two features:   - create a whole MultiTableIncrementer object for each sequence so that caching will be easier when that is implemented   - add the ability to initialize the sequence to the current max value of the identity column

+++ yukon-client/common/src/com/cannontech/database/incrementer/KeyedIncrementer.java
@@ -1,5 +1,5 @@
 package com.cannontech.database.incrementer;
 
 public interface KeyedIncrementer {
-    public int getNextValue(String key);
+    public long getNextValue(String key);
 }
